+++ src/api/middleware/errorHandler.ts
@@ -1,0 +1,6 @@
+import { Request, Response, NextFunction } from 'express';
+
+export default function errorHandler(err: Error, _req: Request, res: Response, _next: NextFunction) {
+  console.error(err.stack);
+  res.status(500).json({ error: err.message || 'Internal server error' });
+}
