+++ src/core/loaders/getStateFromCoordinates.test.ts
