+++ src/core/loaders/getStateFromCoordinates.ts
