+++ src/core/spider/utils/cleanText.test.ts
