+++ src/core/spider/utils/cleanText.ts
